Optionally remove conda-build's test environment
Attempt to remove `conda-build`'s test environment. Though it might not
exist if there was no test step in the recipe. This causes newer
versions of `conda` to fail when we attempt to remove an empty
environment like this. Instead of letting it break the build, 0 out the
exit code from this failure so it merely acts as a warning.