Change the default notebook port to 8888.
This is the default port that the Jupyter Notebook server uses anyways.
So it makes sense to make it the default here. While we don't handle
collisions with existing ports yet, we do provide the ability to change
the port used. So if this conflicts with another Notebook server, it
should be simple to change.