Allow Jupyter Notebook's to be run as `root`
We probably shouldn't be running as `root` honestly. However, it is
inside a container. So we shouldn't be too worried about this. Just a
reminder from other developers that we should start migrating to a
non-`root` user. In the interim, we use the easy fix to get around this
new behavior of the Jupyter Notebook server.