Check the Docker image and tag are present
Previously only the image name was checked in the Docker image list.
However this doesn't validate the specified tag is present. Nor does it
handle cases where the image's tag is also specified with the image
name. This corrects that issue by separating the image name from the
image tag (if there is one) using `latest` otherwise and checking that
both are found together in the image listing.