Drop installation of the `ipyparallel` extension
As `ipyparallel` is no longer used to startup the cluster used for
computations, there is no real need to special configuration of
`ipyparallel`. Further as `ipyparallel` was always started and shutdown
within the Jupyter Notebook, there is no real benefit to having a GUI to
manage this step even for older variants of the workflow (though one
could run this in their own install trivially). After all the GUI is
only used to start and stop the `ipyparallel` cluster not manage it in
any way. Given all of this, we drop installing the `ipyparallel`
extension to the notebook.