Enable `execute_time` in system prefix
When calling `jupyter nbextension enable`, it by default will enable the
extension in the home directory. However this is problematic when trying
to run the Jupyter Notebook in a singularity container as it lacks
permissions to `root`'s home directory. To fix this issue, we add the
`--sys-prefix` argument. This seems to have the correct behavior of
enabling the `execute_time` extension in the `PREFIX/etc` under the
current `conda` install.