Remove the `.conda` directory in `root`'s home
When installing and upgrading packages with `conda`, it creates a
`.conda` directory in `root`'s home. This is used by `conda` as a
fallback if it lacks permissions to `conda`'s install prefix. However
leaving `root`'s `.conda` directory could mess with a user created one
while running the container. So this cleans up the `.conda` directory in
`root`'s home after all of the installing and upgrading is complete.
This should make the container more usable with Singularity.