Remove `.cache` directory from `root`'s home
Appears that running `pip install` on the cloned workflow causes `pip`
to create a `.cache` directory in `root`'s home with some content. As we
are not really interested in this content or keeping it around in the
resulting Docker container of the workflow, make sure to remove the
`.cache` directory from `root`'s home directory before committing the
layer installing the workflow. This should make the container more
usable with Singularity.