Install the workflows in the entrypoint script
To simplify things a bit for other deployment modes (e.g. the cluster),
install the workflows to the current working directory in the
entrypoint script. This should allow users to get started with the
workflow easily in their directory of choosing just as part of startup.
Thus it should be possible to eliminate this from the startup script
used in local development.