Pin distributed to 1.21.0
Currently we are affected by a `distributed` bug in version 1.21.0 on
the cluster, which causes workers to not be registered correctly. This
results in workers terminating early and resources not being correctly
determined. While the resolution to this issue is being pursued, pin
`distributed` to the last known working version (1.21.0) in the Docker
image build. This should produce an image that works as expected on the
cluster for the interim.