Drop unneeded Jupyter extension install steps
Previously we had manually run extension install steps (after doing the
`conda install` of their packages where the post-link steps should have
handled this issue). However the `conda` packages with a modern version
of `conda` seem to behave fine. Plus this is done upstream of the
workflow container in the notebook container. So there is no need to
explicitly run these configuration steps in the workflow container.
Should cutdown noise at this level both in logs and in build layers
(generally simplifying maintenance).