Clean the entire `conda-bld` directory
As the `conda-bld/work` directory is a bit of legacy location from
`conda-build` 1, cleaning it doesn't make lots of sense. Besides we
don't need to keep anything in the `conda-bld` directory as it is just
build remnants. So go and clean the whole `conda-bld` directory
returning it to it's original state.