Unpin Bokeh
Newer versions of Distributed work with Bokeh 0.13.0+ and in fact
require Bokeh 0.13.0 at a minimum. Further newer versions of Dask
require a newer version of Distributed. So requiring an older version of
Bokeh also holds back Dask and Distributed. Given the scheduler display
bug has been fixed with Bokeh 0.13.0 in Distributed 1.23.0, go ahead and
release this Bokeh pin. Thus also allowing newer versions of Dask and
Distributed to be installed.