Force Dask Distributed to use spawn
Sets an environment variable to ensure that Dask Distributed uses spawn
to start workers. This seems to avoid some segmentation faults that
plagued the dictionary learning step when OpenBLAS was running. As
forking and OpenBLAS have had problems in the past, staying clear of
forking for that step should avoid some issues.